refactor(gateway): consume shared OpenAI protocol
