+++ wmo/runtime/gateway/response.py
@@ -16,8 +16,8 @@ from wmo.runtime.gateway.contracts import (
     GatewayRequest,
     GatewayUsage,
 )
-from wmo.runtime.gateway.openai.errors import OpenAIProtocolError
-from wmo.runtime.gateway.openai.streaming import (
+from wmo.runtime.openai_protocol.errors import OpenAIProtocolError
+from wmo.runtime.openai_protocol.streaming import (
     ChatSseEncoder,
     ResponsesSseEncoder,
     stable_public_id,
